Refactored out some components and styles.
