+++ assets/css/elevator_v2.scss
@@ -1,3 +1,7 @@
+// Our Mimo elevator screens run Webview 83.
+// When writing styles for widgets on these screens,
+// verify browser compatibility in the MDN web docs.
+
 @import "https://rsms.me/inter/inter.css";
 @import "colors";
 @import "v2/lcd_common/screen_container";
