+++ assets/css/v2/elevator/closures/current_elevator_closed_view.scss
@@ -1,0 +1,126 @@
+.current-elevator-closed-view {
+  position: relative;
+  height: 100%;
+
+  .shape {
+    position: absolute;
+    top: 0;
+    right: 0;
+    display: inline-block;
+    border-right: 345px solid $accessibility-blue;
+    border-bottom: 300px solid transparent;
+  }
+
+  .header {
+    box-sizing: border-box;
+    height: 528px;
+    padding: 53px 68px 51px 48px;
+    font-weight: 700;
+
+    .icons {
+      margin-bottom: 21px;
+
+      .no-service-icon {
+        margin-right: 16px;
+      }
+    }
+
+    .closed-text {
+      margin-bottom: 18px;
+      font-size: 150px;
+      line-height: 150px;
+    }
+  }
+
+  hr.thin {
+    margin: 0 20px;
+  }
+
+  .accessible-path-container {
+    box-sizing: border-box;
+    height: 984px;
+    padding: 48px;
+
+    .subheading-container {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      height: 142px;
+
+      .subheading {
+        align-self: center;
+      }
+
+      .arrow {
+        width: 100px;
+        height: 100px;
+      }
+    }
+
+    .alternate-direction-text {
+      color: $cool-black-30;
+
+      &.small {
+        font-size: 48px;
+        line-height: 64px;
+      }
+
+      &.medium {
+        font-size: 62px;
+        line-height: 80px;
+      }
+
+      &.large {
+        font-size: 80px;
+        line-height: 96px;
+      }
+    }
+
+    .map-container {
+      position: relative;
+      margin-top: 26px;
+      margin-left: -48px;
+
+      .map {
+        width: 1080px;
+      }
+    }
+  }
+
+  .paging-indicators {
+    justify-content: end;
+    height: 120px;
+  }
+}
+
+.marker-container {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 116px;
+  height: 116px;
+
+  .marker-background,
+  .marker {
+    position: absolute;
+  }
+
+  .marker-background {
+    animation: pulse-ring 1.5s infinite;
+  }
+}
+
+@keyframes pulse-ring {
+  0% {
+    transform: scale(0.5);
+  }
+
+  60% {
+    opacity: 0.25;
+  }
+
+  100% {
+    opacity: 0;
+  }
+}
