+++ assets/css/v2/elevator/closures/outside_closures_view.scss
@@ -1,0 +1,123 @@
+.outside-closures-view {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+
+  .in-station-summary {
+    display: flex;
+    justify-content: space-between;
+    padding: 24px 48px;
+    font-size: 48px;
+    font-weight: 400;
+    line-height: 64px;
+    color: $cool-black-30;
+
+    .text {
+      margin-right: 82px;
+    }
+  }
+
+  .outside-closure-list {
+    height: 100%;
+    background-color: $warm-neutral-90;
+
+    .header-container {
+      margin: 48px;
+      margin-bottom: 0;
+
+      .header {
+        display: flex;
+        max-height: 432px;
+        margin-bottom: 48px;
+        font-size: 112px;
+        font-weight: 700;
+        line-height: 112px;
+
+        &__title {
+          word-spacing: 9999px;
+        }
+      }
+    }
+
+    .closure-list-container {
+      overflow: hidden;
+
+      .closure-list {
+        display: flex;
+        flex-flow: column wrap;
+        height: 904px;
+        transform: translateX(calc(-100% * var(--closure-list-offset)));
+
+        .closure-row {
+          box-sizing: border-box;
+          width: 1080px;
+          padding: 0 48px;
+
+          &__station-name {
+            font-size: 62px;
+            font-weight: 600;
+            line-height: 80px;
+            color: $cool-black-15;
+          }
+
+          &__name-and-pills {
+            display: flex;
+            align-items: center;
+            margin-bottom: 14px;
+
+            .route-pill {
+              width: 132px;
+              height: 68.13px;
+              margin-right: 24px;
+
+              &__text {
+                line-height: 68.13px;
+              }
+
+              &__icon {
+                height: 100%;
+              }
+
+              &__icon-image {
+                height: 100%;
+                margin-top: 0;
+              }
+            }
+          }
+
+          &__elevator-name {
+            line-height: 64px;
+          }
+
+          &__elevator-name.list-item {
+            display: list-item;
+            margin-bottom: 8px;
+            margin-left: 48px;
+          }
+        }
+      }
+    }
+
+    .paging-info-container {
+      position: absolute;
+      bottom: 0;
+      display: flex;
+      justify-content: space-between;
+      width: 100%;
+      height: 72px;
+      margin-bottom: 20px;
+
+      .more-elevators-text {
+        align-self: center;
+        padding: 0 48px 20px;
+      }
+    }
+
+    .paging-info-container,
+    .closure-row__elevator-name {
+      font-size: 48px;
+      font-weight: 400;
+      color: $cool-black-30;
+    }
+  }
+}
